"check_style_score -> 1 of 3 passed. style score: 0/1 "

+++ ADS-1-Assignments/m16/16.1 BST/Solution.java
@@ -14,7 +14,7 @@ public class Solution {
 				break;
 			case "get":
 				key = new Book(tokens[1], tokens[2],
-				               Integer.parseInt(tokens[3]));
+				               Float.parseFloat(tokens[3]));
 				System.out.println(bst.get(key));
 				break;
 			default:
